+++ source/code/CookieEngine/Engine/Entities/EntityAdmin.h
@@ -32,7 +32,7 @@ namespace Cookie {
 		// -----------------------------------------------------------------
 
 		template <typename T> void RegisterComponent() {
-			u32 typeID = typeid(T).hash_code();
+			size_t typeID = typeid(T).hash_code();
 			ComponentArray<T> *array = new ComponentArray<T>();
 			m_ComponentArrays.insert({typeID, array});
 			m_ComponentSignatureIndex.insert({typeID, m_NextComponentIndex});
@@ -62,7 +62,7 @@ namespace Cookie {
 		template <typename T> T *GetComponent(EntityID entity) { return GetComponentArray<T>()->Get(entity); };
 
 		template <typename T> ComponentSignatureIndex GetComponentSignatureID() {
-			u32 typeID = typeid(T).hash_code();
+			size_t typeID = typeid(T).hash_code();
 			CKE_ASSERT(m_ComponentSignatureIndex.find(typeID) != m_ComponentSignatureIndex.end(),
 					   "A system is trying to acess a component that is not registered");
 			return m_ComponentSignatureIndex[typeID];
@@ -72,14 +72,14 @@ namespace Cookie {
 		// -----------------------------------------------------------------
 
 		template <typename T> void RegisterSystem() {
-			u32 typeID = typeid(T).hash_code();
+			size_t typeID = typeid(T).hash_code();
 			System *system = new T();
 			m_Systems.insert({typeID, system});
 			system->InitSignature();
 		};
 
 		template <typename T> void RegisterSystem(T *system) {
-			u32 typeID = typeid(T).hash_code();
+			size_t typeID = typeid(T).hash_code();
 			m_Systems.insert({typeID, system});
 			system->InitSignature();
 		};
@@ -89,11 +89,11 @@ namespace Cookie {
 		TArray<Signature, MAX_ENTITIES> m_Signatures{};
 		u32 m_ActiveEntitiesCount{};
 
-		THashMap<u32, IComponentArray *> m_ComponentArrays{};
-		THashMap<u32, ComponentSignatureIndex> m_ComponentSignatureIndex{};
+		THashMap<size_t, IComponentArray *> m_ComponentArrays{};
+		THashMap<size_t, ComponentSignatureIndex> m_ComponentSignatureIndex{};
 		ComponentSignatureIndex m_NextComponentIndex{};
 
-		THashMap<u32, System *> m_Systems{};
+		THashMap<size_t, System *> m_Systems{};
 
 		// -----------------------------------------------------------------
 
@@ -101,7 +101,7 @@ namespace Cookie {
 			Signature entitySignature = m_Signatures[entityID];
 
 			for (auto const &pair : m_Systems) {
-				u32 systemTypeID = pair.first;
+				size_t systemTypeID = pair.first;
 				System *system = pair.second;
 
 				if ((entitySignature & system->m_Signature) == system->m_Signature) {
@@ -115,7 +115,7 @@ namespace Cookie {
 		}
 
 		template <typename T> ComponentArray<T> *GetComponentArray() {
-			u32 typeID = typeid(T).hash_code();
+			size_t typeID = typeid(T).hash_code();
 			CKE_ASSERT(m_ComponentArrays.find(typeID) != m_ComponentArrays.end(), "Trying to add an unregistered component to an entity");
 			return dynamic_cast<ComponentArray<T> *>(m_ComponentArrays[typeID]);
 		}
